refactor: mainlightcontroller
