+++ app/resources/shaders/basic.glsl
@@ -63,6 +63,7 @@ struct SpotLight {
 
 uniform PointLight pointLightSun;
 uniform PointLight pointLightMoon;
+uniform PointLight pointLightPyramid;
 uniform SpotLight spotLight;
 uniform vec3 viewPosition;
 
@@ -72,8 +73,9 @@ vec3 CalcSpotLight(SpotLight light, vec3 normal, vec3 fragPos, vec3 viewDir);
 void main() {
     vec3 normal = normalize(Normal);
     vec3 viewDir = normalize(viewPosition - FragPos);
-    vec3 result = CalcPointLight(pointLightMoon, normal, FragPos, viewDir);
+    vec3 result = CalcPointLight(pointLightPyramid, normal, FragPos, viewDir);
     result += CalcPointLight(pointLightSun, normal, FragPos, viewDir);
+    result += CalcPointLight(pointLightMoon, normal, FragPos, viewDir);
     result += CalcSpotLight(spotLight, normal, FragPos, viewDir);
     float brightness = dot(result, vec3(0.2126, 0.7152, 0.0722));
     if (brightness > 1.0)
