+++ app/resources/shaders/bloom_final.glsl
@@ -20,13 +20,16 @@ in vec2 TexCoords;
 uniform sampler2D scene;
 uniform sampler2D bloomBlur;
 uniform float exposure;
+uniform bool bloom;
 
 void main()
 {
     const float gamma = 2.2;
     vec3 hdrColor = texture(scene, TexCoords).rgb;
     vec3 bloomColor = texture(bloomBlur, TexCoords).rgb;
-    hdrColor += bloomColor;
+    if (bloom) {
+        hdrColor += bloomColor;
+    }
     vec3 result = vec3(1.0) - exp(-hdrColor * exposure);
     result = pow(result, vec3(1.0 / gamma));
     FragColor = vec4(result, 1.0);
